# investment-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 投资管理系统统一接口契约
  version: 1.1.0
  description: |
    LTC 投资管理系统正式接口契约。认证使用 HttpOnly Cookie，会话写操作同时校验 X-CSRF-Token。
    所有投资业务单据统一按“经办人、经办部门负责人、经办部门分管领导”顺序流转。
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: 本地及部署环境 API 基地址
security:
  - cookieAuth: []
...
paths:
...
  /auth/change-password:
    post:
      operationId: changePassword
      summary: 修改当前用户密码并撤销已有会话
      x-permissions: [authenticated]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordInput'
      responses: *okResponses
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ltc_session
  parameters:
...
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 64, maxLength: 64 }
...
  schemas:
...
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, maxLength: 128, format: password }
        newPassword: { type: string, minLength: 12, maxLength: 128, format: password }
...